// frontend/magic-web/src/opensource/pages/superMagic/pages/Workspace/types.ts
export enum TopicMode {
	/** 通用模式 */
	General = "general",
	/** 聊天模式 */
	Chat = "chat",
	/** 数据分析 */
	DataAnalysis = "data_analysis",
	/** PPT */
	PPT = "ppt",
	/** 研报模式 */
	Report = "report",
	/** 录音总结 */
	RecordSummary = "summary",
	/** 空模式 */
	Empty = "",
	/** 设计模式 */
	Design = "design",
	/** 员工主导模式
	 * 不会在员工列表中存在，只用于创建员工时使用
	 */
	AgentManager = "agent-manager",
}
// ...
export enum TaskStatus {
	WAITING = "waiting",
	RUNNING = "running",
	FINISHED = "finished",
	SUSPENDED = "suspended",
	ERROR = "error",
}

/** 工作区状态 */
export enum WorkspaceStatus {
	/** 未运行 */
	WAITING = "waiting",
	/** 运行中 */
	RUNNING = "running",
}

/** 项目状态 */
export enum ProjectStatus {
	/** 未运行 */
	WAITING = "waiting",
	/** 运行中 */
	RUNNING = "running",
}

// 工作区
export interface Workspace {
	id: string
	name: string
	is_archived: 0 | 1 // 0: 未归档，1: 已归档
	current_topic_id: string
	current_project_id: string | null
	workspace_status: WorkspaceStatus
	project_count: number
	status?: "waiting" | "running" | "finished" | "error" // 其实没有这个字段，但是为了兼容旧数据，所以加了这个字段
}

// 话题
export interface Topic {
	id: string
	user_id: string
	chat_topic_id: string
	chat_conversation_id: string
	topic_name: string
	task_status: TaskStatus
	task_mode: string
	project_id: string
	topic_mode: TopicMode
	updated_at: string
	workspace_id: string
	/** Token usage count. null means do not show usage indicator. */
	token_used: number | null
}
// ...
export interface ProjectListItem {
	id: string
	project_status: ProjectStatus
	project_mode: TopicMode | ""
	project_description?: string // 协作项目才有
	workspace_id: string
	work_dir: string
	workspace_name: string
	project_name: string
	current_topic_id: string
	current_topic_status: string
	user_id?: string // 自己的项目才有
	user_organization_code?: string // 自己的项目才有
	created_uid?: string // 自己的项目才有
	updated_uid?: string // 自己的项目才有
	created_at: string
	updated_at: string
	members?: Collaborator[] // 协作项目才有
	member_count?: number // 协作项目才有
	creator?: CollaborationProjectListItem["creator"] // 协作项目才有
	tag: "collaboration" | "" // 用于标识是否是协作项目
	is_pinned?: boolean // 协作项目才有
	last_active_at?: string // 协作项目才有
	user_role?: CollaboratorPermission
	is_bind_workspace?: boolean // 协作项目才有
	bind_workspace_id?: string // 协作项目才有
}

// 创建新项目的响应数据
export interface CreatedProject {
	project: ProjectListItem
	topic: Topic
}

// 项目协作成员
export interface Collaborator {
	id: string
	name: string
	i18n_name: string
	organization_code: string
	avatar_url: string
	type: "User" | "Department"
	user_id?: string
	department_id?: string
	path_nodes: PathNode[]
	role: CollaboratorPermission
	join_method?: CollaborationJoinMethod
}

export interface CollaborationProjectCreator {
	id: string
	user_id: string
	name: string
	avatar_url: string
}

// 项目协作成员列表
export interface CollaborationProjectListItem {
	id: string
	project_status: ProjectStatus
	workspace_id: string
	project_description: string
	work_dir: string
	current_topic_id: string
	current_topic_status: string
	project_mode: TopicMode | ""
	workspace_name: string
	project_name: string
	created_at: string
	updated_at: string
	creator: Omit<CollaborationProjectCreator, "id" | "name"> & { nickname: string }
	members: Collaborator[]
	member_count: number
	tag: "collaboration"
	is_pinned?: boolean
	last_active_at?: string
	is_bind_workspace: boolean
	bind_workspace_id: string
	user_role?: CollaboratorPermission
}

export const enum CollaborationProjectType {
	/**
	 * 自己分享的项目
	 */
	Shared = "shared",
	/**
	 * 别人分享给自己的项目
	 */
	Received = "received",
}

/**
 * 协作邀请方式
 */
export const enum CollaborationJoinMethod {
	Internal = "internal",
	link = "link",
}